Add chrReadHpResult Structure and Implement Logging Methods
- Introduce the `chrReadHpResult` structure to encapsulate read HP results and related methods.
- Implement `recordReadHp`, `recordDeriveHp`, and `recordAlignCoverRegion` methods for logging read HP data, derived HP, and coverage region updates.
- Remove the old implementation of these methods from `HaplotagType.cpp` to streamline the codebase.
- Update header files to reflect the new structure and method definitions, enhancing overall organization and clarity.
